Make OS specific Tor controlling
